// bar-frontend/src/app/api-client/model/itemDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CategoryDto } from './categoryDto';
import { ItemTagDto } from './itemTagDto';


export interface ItemDto { 
    id?: number;
    createdBy?: string;
    lastModifiedBy?: string;
    createdDate?: string;
    lastModifiedDate?: string;
    name?: string;
    description?: string;
    price?: number;
    photoUrl?: string;
    tag?: ItemTagDto;
    category?: CategoryDto;
}